comment out applescript classes handling
With the built in headerdoc, I was getting the error
'"headerdoc" Can't call method "classes" on an undefined value'.

I have commented out the part that deals with applescript classes
and that seems to fix the problem. I'm not sure what this code was
supposed to be doing, or if this will work for more complicated usage,
but commenting this out worked for me and my needs.